+++ src/TextureStudio.App/Program.cs
@@ -23,6 +23,7 @@ builder.Services.AddScoped<OpenAiImageClient>();
 builder.Services.AddScoped<ImageCodec>();
 builder.Services.AddScoped<WorkspaceService>();
 builder.Services.AddScoped<ContentSearchService>();
+builder.Services.AddScoped<BrowserSupport>();
 builder.Services.AddScoped<StudioState>();
 
 await builder.Build().RunAsync();